+++ CHANGELOG.md
@@ -20,7 +20,8 @@ All notable changes to this project should be documented in this file.
 - created `src/test/systemtests/test_scorm.py` that tests the scormbridge in particular and scorm communication in general
 - moved lots of functionalities from newIntersite.js to other more fitting modules
   - pushIso: now uses scormBridge to update points
-  - 
+  - some helper functions to new js module `veHelpers.js`
+- 
 
 ### Fixed
 - user input is now reloaded into input fields again #68